+++ lib/service/Assistant.dart
@@ -21,6 +21,7 @@ class Assistant with Service, NotificationsListener {
 
   AssistantUser? _user;
   AssistantSettings? _settings;
+  String? _sessionID;
 
   DateTime?  _pausedDateTime;
 
@@ -62,7 +63,7 @@ class Assistant with Service, NotificationsListener {
 
   @override
   Future<void> initService() async {
-
+    _sessionID = Storage().assistantSessionID;
     List<Future<String?>> initFutures = <Future<String?>>[];
 
     int? futuresSettingsIndex;
@@ -278,8 +279,14 @@ class Assistant with Service, NotificationsListener {
     }
   }
 
+  Future<void> resetSession() async {
+    Storage().assistantSessionID = _sessionID = null;
+  }
+
   Future<bool> removeAllMessages() async {
     bool succeeded = false;
+    // remove session id here
+    resetSession();
     if (_isEnabled) {
       String url = '${Config().aiProxyUrl}/messages';
       Response? response = await Network().delete(url, auth: Auth2());
@@ -349,6 +356,10 @@ class Assistant with Service, NotificationsListener {
       };
     }
 
+    if (_sessionID != null) {
+      body['session_id'] = _sessionID;
+    }
+
     try {
       String? json = JsonUtils.encode(body);
       Response? response = await Network().post(url, auth: Auth2(), headers: headers, body: json);
@@ -357,7 +368,13 @@ class Assistant with Service, NotificationsListener {
       if (responseCode == 200) {
         Map<String, dynamic>? responseJson = JsonUtils.decodeMap(responseString);
         if (responseJson != null) {
-          return Message.fromAnswerJson(responseJson);
+          //get and update session id
+          Message message = Message.fromAnswerJson(responseJson);
+
+          if (message.modelResponseId != null) {
+            Storage().assistantSessionID = _sessionID = message.modelResponseId;
+          }
+          return message;
         }
         return null;
       } else {